Fixes issue with tooltips and scrolling.
Before this commit, if you had a chart with tooltips enabled and
had to scroll down to get to the chart, the tooltip would appear
off screen, or much higher than it should have because the scroll
value was not considered.

This fix ensures that the scroll is taken into account whenever
the tooltip position is calculated.